LayoutTests/fast/shadow-dom/mouseenter-mouseleave-on-slot-parent-expected.txt

This tests mouseenter and mouseleave event fires when capturing event listeners are only present on the slot element's parent.
To manually test, move the mouse cursor into the box below and move it out.
You should see mouseenter events on slotContainer, slot, and target in that order and mouseleave events in the reverse order.


==Entering target==
mouseenter on target
mouseenter on slot
mouseenter on slotContainer

==Leaving target==
mouseleave on target
mouseleave on slot
mouseleave on slotContainer
